requirements.txt
setuptools
viam-sdk
reolinkapi @ git+https://github.com/stevebriskin/reolinkapipy.git
typing-extensions
Pillow
reolinkapi[streaming] @ git+https://github.com/stevebriskin/reolinkapipy.git
PyInstaller
